pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mongodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["python-dotenv>=1.0.0", "pymongo>=4.10.0", "streamlit>=1.30.0"]

[tool.setuptools.packages.find]
include = ["dataset*", "view*"]
